Assignment-2/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.4
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
certifi==2026.2.25
cfgv==3.5.0
charset-normalizer==3.4.6
click==8.3.1
cuda-bindings==13.2.0
cuda-pathfinder==1.5.0
cuda-toolkit==13.0.2
dataclasses-json==0.6.7
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
docstring_parser==0.17.0
faiss-cpu==1.13.2
filelock==3.25.2
frozenlist==1.8.0
fsspec==2026.3.0
greenlet==3.3.2
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.8.0
identify==2.6.18
idna==3.11
iniconfig==2.3.0
instructor==1.14.5
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.1.1
langchain==1.2.13
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.23
langchain-huggingface==1.2.1
langchain-text-splitters==1.1.1
langgraph==1.1.3
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langsmith==0.7.22
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
networkx==3.6.1
nodeenv==1.10.0
numpy==2.4.4
nvidia-cublas==13.1.0.3
nvidia-cuda-cupti==13.0.85
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-runtime==13.0.96
nvidia-cudnn-cu13==9.19.0.56
nvidia-cufft==12.0.0.61
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-cusolver==12.0.4.66
nvidia-cusparse==12.6.3.3
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
openai==2.30.0
openai-swarm==0.1.1
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
platformdirs==4.9.4
pluggy==1.6.0
pre_commit==4.5.1
propcache==0.4.1
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.20.0
pypdf==6.9.2
pytest==9.0.2
python-discovery==1.2.1
python-dotenv==1.2.2
PyYAML==6.0.3
regex==2026.3.32
requests==2.33.0
requests-toolbelt==1.0.0
rich==14.3.3
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.3.0
setuptools==81.0.0
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.48
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.11.0
tqdm==4.67.3
transformers==5.4.0
triton==3.6.0
typer==0.24.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
virtualenv==21.2.0
xxhash==3.6.0
yarl==1.23.0
zstandard==0.25.0
